+++ src/components/flyout/flyout.tsx
@@ -55,8 +55,7 @@ export interface EuiFlyoutProps
    */
   hideCloseButton?: boolean;
   /**
-   * Locks the mouse / keyboard focus to within the flyout,
-   * and shows an EuiOverlayMask
+   * Adds an EuiOverlayMask when set to `true`
    */
   ownFocus?: boolean;
   /**
@@ -172,9 +171,17 @@ export const EuiFlyout: FunctionComponent<EuiFlyoutProps> = ({
     <Fragment>
       <EuiWindowEvent event="keydown" handler={onKeyDown} />
       {optionalOverlay}
-      {/* Trap focus even when ownFocus={false}, otherwise closing the flyout won't return focus
-        to the originating button */}
-      <EuiFocusTrap clickOutsideDisables={false}>{flyoutContent}</EuiFocusTrap>
+      {/*
+       * Trap focus even when `ownFocus={false}`, otherwise closing
+       * the flyout won't return focus to the originating button.
+       *
+       * Set `clickOutsideDisables={true}` when `ownFocus={false}`
+       * to allow non-keyboard users the ability to interact with
+       * elements outside the flyout.
+       */}
+      <EuiFocusTrap clickOutsideDisables={!ownFocus}>
+        {flyoutContent}
+      </EuiFocusTrap>
     </Fragment>
   );
 };